Break out worker list sync, add v2 docker-compose
This modernizes the use of docker-compose (avoiding the need for links
and using structured labels) and adds a "config" container to listen
for Docker events, update the worker list file, and SIGHUP the master
to reload configuration files.

This removes the need for hardcoding the number of workers, worker
ports, links, and the reload-workers shell script.